rename get_new_tasks -> get_new_suggestions_or_automerge
